Renamed apis and added more error handling for data deserialisation
